requirements.txt
asgiref==3.5.2
astroid==2.11.6
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
certifi==2022.6.15
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.0.12
commonmark==0.9.1
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==38.0.3
daphne==4.0.0
defusedxml==0.7.1
dill==0.3.5.1
dj-database-url==0.5.0
Django==4.1.4
django-ajax-selects==2.2.0
django-appconf==1.0.5
django-bootstrap==0.2.4
django-bootstrap5==22.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-dotenv==1.4.2
django-environ==0.9.0
django-filter==22.1
django-mailer==2.2
django-model-utils==4.3.1
django-phone-field==1.8.1
django-phonenumber-field==6.3.0
django-private-chat2==1.0.2
django-registration==3.3
djangorestframework==3.13.1
exceptiongroup==1.0.4
gunicorn==20.1.0
h11==0.14.0
hashring==1.5.1
hiredis==2.0.0
httpie==3.2.1
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
isort==5.10.1
lazy-object-proxy==1.7.1
lockfile==0.12.2
Markdown==3.3.7
mccabe==0.7.0
memcache==0.12.0
msgpack==1.0.4
multidict==6.0.2
outcome==1.2.0
packaging==21.3
phonenumbers==8.12.50
Pillow==9.1.1
pinax-notifications==6.0.0
platformdirs==2.5.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pylint==2.14.3
pymemcache==4.0.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.1
redis==4.3.5
requests==2.28.0
requests-toolbelt==0.9.1
rich==12.4.4
selenium==4.7.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.9
whitenoise==6.2.0
wrapt==1.14.1
wsproto==1.2.0
zope.interface==5.5.2
